add `ui/window.rs` module
move all window rendering functions into `window.rs`
